AI/mcp-server/mcp-leave-manager/leave-manager-server/pyproject.toml
[project]
name = "leave-manager-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "mcp[cli]>=1.14.0",
    "psycopg2-binary>=2.9.10",
]
